+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * For players:
+  * Added friendly error when using SMAPI 3.2._x_ with Stardew Valley 1.5.5 or later.
   * Improved mod compatibility in 64-bit mode (thanks to spacechase0!).
   * Reducing load time when scanning/rewriting many mods for compatibility.
   * Updated compatibility list.